Refactor: Update component imports and remove deprecated files
- Removed unused components: LoggingTest, GlobalStateDemo, and their associated styles.
- Updated imports in various files to reflect the new structure, including moving FormInput and FormSelect to adapters.
- Simplified App.js by removing LoggingTest and GlobalStateDemo routes.
- Adjusted button components across admin pages to use the new Button component from the UI library.
- Cleaned up the components directory by deleting old UI components that are now replaced by Mantine-based adapters.
