+++ frontend/src/components/auth/AdminResetPasswordModal.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Modal from '../ui/Modal';
+import { Modal } from '../ui';
 import { Button, Alert } from '../ui';
 import { adminApiService } from '../../services/api/adminApi';
