--- frontend/src/components/common/GlobalStateDemo.css
@@ -1,189 +1,0 @@
-.global-state-demo {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 2rem;
-  font-family:
-    -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
-}
-
-.global-state-demo h2 {
-  color: #2c3e50;
-  margin-bottom: 1rem;
-  text-align: center;
-  border-bottom: 2px solid #3498db;
-  padding-bottom: 0.5rem;
-}
-
-.demo-description {
-  background: #e8f4fd;
-  border: 2px solid #3498db;
-  border-radius: 8px;
-  padding: 1rem;
-  margin-bottom: 2rem;
-  color: #2c3e50;
-  font-size: 1.1rem;
-  text-align: center;
-}
-
-.cache-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(300px, 1fr));
-  gap: 1.5rem;
-  margin-bottom: 2rem;
-}
-
-.cache-item {
-  background: #ffffff;
-  border: 2px solid #ecf0f1;
-  border-radius: 12px;
-  padding: 1.5rem;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  transition: all 0.3s ease;
-}
-
-.cache-item:hover {
-  transform: translateY(-2px);
-  box-shadow: 0 6px 12px rgba(0, 0, 0, 0.15);
-  border-color: #3498db;
-}
-
-.cache-item h3 {
-  color: #2c3e50;
-  margin: 0 0 1rem 0;
-  font-size: 1.3rem;
-  border-bottom: 1px solid #bdc3c7;
-  padding-bottom: 0.5rem;
-}
-
-.cache-status {
-  margin-bottom: 1rem;
-}
-
-.global-state-demo .status-badge {
-  display: inline-block;
-  padding: 0.4rem 0.8rem;
-  border-radius: 20px;
-  font-size: 0.9rem;
-  font-weight: 600;
-  text-transform: uppercase;
-  letter-spacing: 0.5px;
-}
-
-.status-badge.loading {
-  background: #f39c12;
-  color: white;
-  animation: pulse 1.5s infinite;
-}
-
-.status-badge.ready {
-  background: #27ae60;
-  color: white;
-}
-
-@keyframes pulse {
-  0% {
-    opacity: 1;
-  }
-  50% {
-    opacity: 0.7;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.cache-details p {
-  margin: 0.5rem 0;
-  color: #34495e;
-  font-size: 0.95rem;
-}
-
-.cache-details strong {
-  color: #2c3e50;
-  font-weight: 600;
-}
-
-.cache-controls {
-  background: #f8f9fa;
-  border: 2px solid #dee2e6;
-  border-radius: 12px;
-  padding: 1.5rem;
-  margin-bottom: 2rem;
-  text-align: center;
-}
-
-.cache-controls h3 {
-  color: #2c3e50;
-  margin: 0 0 1rem 0;
-  font-size: 1.2rem;
-}
-
-.cache-button {
-  background: #3498db;
-  color: white;
-  border: none;
-  padding: 0.7rem 1.2rem;
-  margin: 0.3rem;
-  border-radius: 8px;
-  cursor: pointer;
-  font-size: 0.9rem;
-  font-weight: 500;
-  transition: all 0.3s ease;
-  min-width: 180px;
-}
-
-.cache-button:hover {
-  background: #2980b9;
-  transform: translateY(-1px);
-  box-shadow: 0 4px 8px rgba(52, 152, 219, 0.3);
-}
-
-.cache-button:active {
-  transform: translateY(0);
-  box-shadow: 0 2px 4px rgba(52, 152, 219, 0.3);
-}
-
-.cache-settings {
-  background: #f1f2f6;
-  border: 2px solid #a4b0be;
-  border-radius: 12px;
-  padding: 1.5rem;
-}
-
-.cache-settings h3 {
-  color: #2c3e50;
-  margin: 0 0 1rem 0;
-  font-size: 1.2rem;
-}
-
-.cache-settings ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.cache-settings li {
-  background: white;
-  margin: 0.5rem 0;
-  padding: 0.7rem 1rem;
-  border-radius: 6px;
-  border-left: 4px solid #3498db;
-  color: #2c3e50;
-  font-weight: 500;
-}
-
-/* Responsive design */
-@media (max-width: 768px) {
-  .global-state-demo {
-    padding: 1rem;
-  }
-
-  .cache-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .cache-button {
-    width: 100%;
-    margin: 0.3rem 0;
-  }
-}
