--- frontend/src/components/common/LoggingTest.css
@@ -1,254 +1,0 @@
-.logging-test {
-  max-width: 1200px;
-  margin: 0 auto;
-  padding: 20px;
-  font-family:
-    -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', sans-serif;
-}
-
-.logging-test h2 {
-  color: #2c3e50;
-  border-bottom: 2px solid #3498db;
-  padding-bottom: 10px;
-}
-
-.logging-test h3 {
-  color: #34495e;
-  margin-top: 30px;
-  margin-bottom: 15px;
-}
-
-.test-controls {
-  background: #f8f9fa;
-  border: 1px solid #dee2e6;
-  border-radius: 8px;
-  padding: 20px;
-  margin: 20px 0;
-}
-
-.button-grid {
-  display: grid;
-  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  gap: 10px;
-  margin-bottom: 15px;
-}
-
-.utility-buttons {
-  display: flex;
-  gap: 10px;
-  justify-content: center;
-  margin-top: 15px;
-  padding-top: 15px;
-  border-top: 1px solid #dee2e6;
-}
-
-.btn {
-  padding: 10px 15px;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-  font-weight: 500;
-  transition: all 0.2s ease;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 5px;
-}
-
-.btn:hover {
-  transform: translateY(-1px);
-  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
-}
-
-.btn-primary {
-  background-color: #3498db;
-  color: white;
-}
-
-.btn-primary:hover {
-  background-color: #2980b9;
-}
-
-.btn-secondary {
-  background-color: #95a5a6;
-  color: white;
-}
-
-.btn-secondary:hover {
-  background-color: #7f8c8d;
-}
-
-.btn-warning {
-  background-color: #f39c12;
-  color: white;
-}
-
-.btn-warning:hover {
-  background-color: #e67e22;
-}
-
-.btn-danger {
-  background-color: #e74c3c;
-  color: white;
-}
-
-.btn-danger:hover {
-  background-color: #c0392b;
-}
-
-.btn-info {
-  background-color: #17a2b8;
-  color: white;
-}
-
-.btn-info:hover {
-  background-color: #138496;
-}
-
-.btn-error {
-  background-color: #dc3545;
-  color: white;
-}
-
-.btn-error:hover {
-  background-color: #c82333;
-}
-
-.btn-light {
-  background-color: #f8f9fa;
-  color: #495057;
-  border: 1px solid #dee2e6;
-}
-
-.btn-light:hover {
-  background-color: #e2e6ea;
-}
-
-.log-display {
-  background: white;
-  border: 1px solid #dee2e6;
-  border-radius: 8px;
-  padding: 20px;
-  margin: 20px 0;
-}
-
-.log-container {
-  background: #f8f9fa;
-  border: 1px solid #dee2e6;
-  border-radius: 4px;
-  padding: 15px;
-  max-height: 300px;
-  overflow-y: auto;
-  font-family: 'Courier New', monospace;
-  font-size: 14px;
-}
-
-.no-logs {
-  color: #6c757d;
-  font-style: italic;
-  text-align: center;
-  margin: 0;
-}
-
-.log-entry {
-  margin-bottom: 8px;
-  padding: 5px 8px;
-  border-radius: 3px;
-  border-left: 3px solid;
-}
-
-.log-success {
-  background-color: #d4edda;
-  border-left-color: #28a745;
-  color: #155724;
-}
-
-.log-error {
-  background-color: #f8d7da;
-  border-left-color: #dc3545;
-  color: #721c24;
-}
-
-.log-warning {
-  background-color: #fff3cd;
-  border-left-color: #ffc107;
-  color: #856404;
-}
-
-.log-info {
-  background-color: #d1ecf1;
-  border-left-color: #17a2b8;
-  color: #0c5460;
-}
-
-.log-timestamp {
-  color: #6c757d;
-  margin-right: 10px;
-}
-
-.log-message {
-  font-weight: 500;
-}
-
-.instructions {
-  background: #e7f3ff;
-  border: 1px solid #b3d9ff;
-  border-radius: 8px;
-  padding: 20px;
-  margin: 20px 0;
-}
-
-.instructions ol {
-  margin: 10px 0;
-  padding-left: 20px;
-}
-
-.instructions li {
-  margin-bottom: 8px;
-  line-height: 1.5;
-}
-
-.instructions code {
-  background: #f1f3f4;
-  padding: 2px 6px;
-  border-radius: 3px;
-  font-family: 'Courier New', monospace;
-  font-size: 12px;
-}
-
-.status-info {
-  background: #f8f9fa;
-  border: 1px solid #dee2e6;
-  border-radius: 8px;
-  padding: 20px;
-  margin: 20px 0;
-}
-
-.status-info ul {
-  list-style: none;
-  padding: 0;
-  margin: 10px 0;
-}
-
-.status-info li {
-  padding: 5px 0;
-  border-bottom: 1px solid #e9ecef;
-}
-
-.status-info li:last-child {
-  border-bottom: none;
-}
-
-@media (max-width: 768px) {
-  .logging-test {
-    padding: 10px;
-  }
-
-  .button-grid {
-    grid-template-columns: 1fr;
-  }
-
-  .utility-buttons {
-    flex-direction: column;
-  }
-}
