--- frontend/src/components/common/ProtectedRoute.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import { Navigate } from 'react-router-dom';
-
-const ProtectedRoute = ({ children }) => {
-  const token = localStorage.getItem('token');
-
-  if (!token) {
-    return <Navigate to="/login" />;
-  }
-
-  return children;
-};
-
-export default ProtectedRoute;
