+++ frontend/src/components/common/index.js
@@ -1,7 +1,4 @@
 // Common components that are shared across the application
 export { default as ErrorBoundary } from './ErrorBoundary';
-export { default as GlobalStateDemo } from './GlobalStateDemo';
-export { default as LoggingTest } from './LoggingTest';
-export { default as ProtectedRoute } from './ProtectedRoute';
 export { default as GlobalSearch } from './GlobalSearch';
 export { default as SearchResults } from './SearchResults';
