--- frontend/src/components/ui/Button.js
@@ -1,50 +1,0 @@
-/**
- * Common Button component with consistent styling
- */
-
-import React from 'react';
-// CSS now handled by Mantine adapter
-
-const Button = ({
-  children,
-  variant = 'primary',
-  size = 'medium',
-  disabled = false,
-  loading = false,
-  onClick,
-  type = 'button',
-  className = '',
-  ...props
-}) => {
-  const baseClass = 'btn';
-  const variantClass = `btn-${variant}`;
-  const sizeClass = `btn-${size}`;
-  const disabledClass = disabled || loading ? 'btn-disabled' : '';
-  const loadingClass = loading ? 'btn-loading' : '';
-
-  const buttonClass = [
-    baseClass,
-    variantClass,
-    sizeClass,
-    disabledClass,
-    loadingClass,
-    className,
-  ]
-    .filter(Boolean)
-    .join(' ');
-
-  return (
-    <button
-      type={type}
-      className={buttonClass}
-      disabled={disabled || loading}
-      onClick={onClick}
-      {...props}
-    >
-      {loading && <span className="btn-spinner"></span>}
-      {children}
-    </button>
-  );
-};
-
-export default Button;
