--- frontend/src/components/ui/Modal.js
@@ -1,69 +1,0 @@
-/**
- * Modal component for overlays and dialogs
- */
-
-import React, { useEffect } from 'react';
-// CSS now handled by Mantine adapter
-
-const Modal = ({
-  isOpen,
-  onClose,
-  title,
-  children,
-  size = 'medium',
-  showCloseButton = true,
-  closeOnOverlayClick = true,
-  className = '',
-}) => {
-  // Handle escape key
-  useEffect(() => {
-    const handleEscape = e => {
-      if (e.key === 'Escape' && isOpen) {
-        onClose();
-      }
-    };
-
-    if (isOpen) {
-      document.addEventListener('keydown', handleEscape);
-      document.body.style.overflow = 'hidden';
-    }
-
-    return () => {
-      document.removeEventListener('keydown', handleEscape);
-      document.body.style.overflow = 'unset';
-    };
-  }, [isOpen, onClose]);
-
-  if (!isOpen) return null;
-
-  const handleOverlayClick = e => {
-    if (e.target === e.currentTarget && closeOnOverlayClick) {
-      onClose();
-    }
-  };
-
-  return (
-    <div className="modal-overlay" onClick={handleOverlayClick}>
-      <div className={`modal-content modal-${size} ${className}`}>
-        {(title || showCloseButton) && (
-          <div className="modal-header">
-            {title && <h2 className="modal-title">{title}</h2>}
-            {showCloseButton && (
-              <button
-                className="modal-close-btn"
-                onClick={onClose}
-                aria-label="Close modal"
-              >
-                ×
-              </button>
-            )}
-          </div>
-        )}
-
-        <div className="modal-body">{children}</div>
-      </div>
-    </div>
-  );
-};
-
-export default Modal;
