--- frontend/src/styles/pages/ConditionCards.css
@@ -1,41 +1,0 @@
-/* Condition Card Enhancements */
-.condition-icon {
-  margin-right: 8px;
-  font-size: 1.1em;
-}
-
-/* Severity Badge Styles */
-.severity-badge {
-  display: inline-flex;
-  align-items: center;
-  padding: 2px 8px;
-  border-radius: 4px;
-  font-size: 0.875rem;
-  font-weight: 500;
-  text-transform: capitalize;
-}
-
-.severity-mild {
-  background-color: #e8f5e8;
-  color: #2d5a2d;
-  border: 1px solid #c3e6c3;
-}
-
-.severity-moderate {
-  background-color: #fff3cd;
-  color: #856404;
-  border: 1px solid #ffeaa7;
-}
-
-.severity-severe {
-  background-color: #f8d7da;
-  color: #721c24;
-  border: 1px solid #f5c6cb;
-}
-
-.severity-critical {
-  background-color: #f8d7da;
-  color: #721c24;
-  border: 1px solid #f5c6cb;
-  font-weight: 600;
-}
